feat(card-options): add dedicated /settings/card-options page
Extract the card-options form body into a reusable <CardOptionsForm>
so it can be rendered both in SettingsModal (existing contextual use)
and on a new dedicated page at /settings/card-options. The modal now
offers an "Open full page" escape hatch that preserves the current
pathname + query as returnTo, so Back from DefineRules-nested usage
returns to /search (not /upload).
