+++ src/components/modals/SettingsModal/SettingsModal.module.css
@@ -55,3 +55,20 @@
   font-weight: var(--font-semibold);
   margin: 0;
 }
+
+.headerActions {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.openFullPage {
+  font-size: var(--text-xs);
+  color: var(--color-primary);
+  text-decoration: none;
+  white-space: nowrap;
+}
+
+.openFullPage:hover {
+  text-decoration: underline;
+}
